Restructure project into modular package
