+++ auto_reviews_parser/src/analyzers/__init__.py
@@ -1,0 +1,3 @@
+from .data_analyzer import ReviewsAnalyzer
+
+__all__ = ["ReviewsAnalyzer"]
