+++ auto_reviews_parser/src/config/__init__.py
@@ -1,0 +1,3 @@
+from .settings import Config
+
+__all__ = ["Config"]
