+++ auto_reviews_parser/src/database/__init__.py
@@ -1,0 +1,3 @@
+from .reviews_database import ReviewsDatabase
+
+__all__ = ["ReviewsDatabase"]
