+++ auto_reviews_parser/src/models/__init__.py
@@ -1,0 +1,3 @@
+from .review_data import ReviewData
+
+__all__ = ["ReviewData"]
