+++ auto_reviews_parser/src/models/review_data.py
