+++ auto_reviews_parser/src/parsers/__init__.py
@@ -1,6 +1,5 @@
 from .drom_parser import DromParser
 from .drive2_parser import Drive2Parser
 from .base_parser import BaseParser
-from .models import ReviewData
 
-__all__ = ["BaseParser", "DromParser", "Drive2Parser", "ReviewData"]
+__all__ = ["BaseParser", "DromParser", "Drive2Parser"]
