+++ auto_reviews_parser/src/services/__init__.py
@@ -1,0 +1,3 @@
+from .auto_reviews_parser import AutoReviewsParser
+
+__all__ = ["AutoReviewsParser"]
