+++ auto_reviews_parser/src/services/auto_reviews_parser.py
@@ -1,6 +1,4 @@
-#!/usr/bin/env python3
-"""
-Стабильный парсер отзывов и бортжурналов для автомобилей
+"""Стабильный парсер отзывов и бортжурналов для автомобилей
 Собирает данные с Drom.ru и Drive2.ru в базу данных
 Работает в щадящем режиме для долгосрочного сбора данных
 """
@@ -22,285 +20,10 @@ from botasaurus.request import request, Request
 from botasaurus.soupify import soupify
 from botasaurus import bt
 
-# ==================== НАСТРОЙКИ ====================
-
-
-class Config:
-    """Конфигурация парсера"""
-
-    # База данных
-    DB_PATH = "auto_reviews.db"
-
-    # Задержки (в секундах)
-    MIN_DELAY = 5  # Минимальная задержка между запросами
-    MAX_DELAY = 15  # Максимальная задержка между запросами
-    ERROR_DELAY = 30  # Задержка при ошибке
-    RATE_LIMIT_DELAY = 300  # Задержка при rate limit (5 минут)
-
-    # Ограничения
-    MAX_RETRIES = 3  # Максимальное количество повторов
-    PAGES_PER_SESSION = 50  # Страниц за сессию
-    MAX_REVIEWS_PER_MODEL = 1000  # Максимум отзывов на модель
-
-    # User agents для ротации
-    USER_AGENTS = [
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/119.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-    ]
-
-    # Список популярных брендов и моделей для парсинга
-    TARGET_BRANDS = {
-        "toyota": ["camry", "corolla", "rav4", "highlander", "prius", "land-cruiser"],
-        "volkswagen": ["polo", "golf", "passat", "tiguan", "touareg", "jetta"],
-        "nissan": ["qashqai", "x-trail", "almera", "teana", "murano", "pathfinder"],
-        "hyundai": ["solaris", "elantra", "tucson", "santa-fe", "creta", "sonata"],
-        "kia": ["rio", "cerato", "sportage", "sorento", "soul", "optima"],
-        "mazda": ["mazda3", "mazda6", "cx-5", "cx-3", "mx-5", "cx-9"],
-        "ford": ["focus", "fiesta", "mondeo", "kuga", "explorer", "ecosport"],
-        "chevrolet": ["cruze", "aveo", "captiva", "lacetti", "tahoe", "suburban"],
-        "skoda": ["octavia", "rapid", "fabia", "superb", "kodiaq", "karoq"],
-        "renault": ["logan", "sandero", "duster", "kaptur", "megane", "fluence"],
-        "mitsubishi": ["lancer", "outlander", "asx", "pajero", "eclipse-cross", "l200"],
-        "honda": ["civic", "accord", "cr-v", "pilot", "fit", "hr-v"],
-        "bmw": ["3-series", "5-series", "x3", "x5", "x1", "1-series"],
-        "mercedes-benz": ["c-class", "e-class", "s-class", "glc", "gle", "gla"],
-        "audi": ["a3", "a4", "a6", "q3", "q5", "q7"],
-        "lada": ["granta", "kalina", "priora", "vesta", "xray", "largus"],
-    }
-
-
-# ==================== МОДЕЛИ ДАННЫХ ====================
-
-from parsers import ReviewData
-
-# ==================== БАЗА ДАННЫХ ====================
-
-
-class ReviewsDatabase:
-    """Управление базой данных отзывов"""
-
-    def __init__(self, db_path: str = Config.DB_PATH):
-        self.db_path = db_path
-        self.init_database()
-
-    def init_database(self):
-        """Инициализация базы данных"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        # Таблица отзывов
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS reviews (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                source TEXT NOT NULL,
-                type TEXT NOT NULL,
-                brand TEXT NOT NULL,
-                model TEXT NOT NULL,
-                generation TEXT,
-                year INTEGER,
-                url TEXT UNIQUE,
-                title TEXT,
-                content TEXT,
-                author TEXT,
-                rating REAL,
-                pros TEXT,
-                cons TEXT,
-                mileage INTEGER,
-                engine_volume REAL,
-                fuel_type TEXT,
-                transmission TEXT,
-                body_type TEXT,
-                drive_type TEXT,
-                publish_date DATETIME,
-                views_count INTEGER,
-                likes_count INTEGER,
-                comments_count INTEGER,
-                parsed_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                content_hash TEXT UNIQUE,
-                UNIQUE(url, content_hash)
-            )
-        """
-        )
-
-        # Таблица статистики парсинга
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS parsing_stats (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                session_id TEXT,
-                brand TEXT,
-                model TEXT,
-                source TEXT,
-                pages_parsed INTEGER DEFAULT 0,
-                reviews_found INTEGER DEFAULT 0,
-                errors_count INTEGER DEFAULT 0,
-                started_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                finished_at DATETIME,
-                status TEXT DEFAULT 'running'
-            )
-        """
-        )
-
-        # Таблица источников (для отслеживания прогресса)
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS sources_queue (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                brand TEXT,
-                model TEXT,
-                source TEXT,
-                base_url TEXT,
-                priority INTEGER DEFAULT 1,
-                status TEXT DEFAULT 'pending',
-                last_parsed DATETIME,
-                total_pages INTEGER DEFAULT 0,
-                parsed_pages INTEGER DEFAULT 0,
-                created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-            )
-        """
-        )
-
-        # Индексы для быстрого поиска
-        cursor.execute(
-            "CREATE INDEX IF NOT EXISTS idx_brand_model ON reviews(brand, model)"
-        )
-        cursor.execute(
-            "CREATE INDEX IF NOT EXISTS idx_source_type ON reviews(source, type)"
-        )
-        cursor.execute("CREATE INDEX IF NOT EXISTS idx_parsed_at ON reviews(parsed_at)")
-        cursor.execute(
-            "CREATE INDEX IF NOT EXISTS idx_content_hash ON reviews(content_hash)"
-        )
-
-        conn.commit()
-        conn.close()
-
-    def save_review(self, review: ReviewData) -> bool:
-        """Сохранение отзыва в базу"""
-        try:
-            conn = sqlite3.connect(self.db_path)
-            cursor = conn.cursor()
-
-            cursor.execute(
-                """
-                INSERT OR IGNORE INTO reviews (
-                    source, type, brand, model, generation, year, url, title, content,
-                    author, rating, pros, cons, mileage, engine_volume, fuel_type,
-                    transmission, body_type, drive_type, publish_date, views_count,
-                    likes_count, comments_count, parsed_at, content_hash
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-                (
-                    review.source,
-                    review.type,
-                    review.brand,
-                    review.model,
-                    review.generation,
-                    review.year,
-                    review.url,
-                    review.title,
-                    review.content,
-                    review.author,
-                    review.rating,
-                    review.pros,
-                    review.cons,
-                    review.mileage,
-                    review.engine_volume,
-                    review.fuel_type,
-                    review.transmission,
-                    review.body_type,
-                    review.drive_type,
-                    review.publish_date,
-                    review.views_count,
-                    review.likes_count,
-                    review.comments_count,
-                    review.parsed_at,
-                    review.content_hash,
-                ),
-            )
-
-            success = cursor.rowcount > 0
-            conn.commit()
-            conn.close()
-            return success
-
-        except sqlite3.IntegrityError:
-            # Дублирующая запись
-            return False
-        except Exception as e:
-            logging.error(f"Ошибка сохранения отзыва: {e}")
-            return False
-
-    def get_reviews_count(self, brand: str = None, model: str = None) -> int:
-        """Получение количества отзывов"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        if brand and model:
-            cursor.execute(
-                "SELECT COUNT(*) FROM reviews WHERE brand = ? AND model = ?",
-                (brand, model),
-            )
-        elif brand:
-            cursor.execute("SELECT COUNT(*) FROM reviews WHERE brand = ?", (brand,))
-        else:
-            cursor.execute("SELECT COUNT(*) FROM reviews")
-
-        count = cursor.fetchone()[0]
-        conn.close()
-        return count
-
-    def is_url_parsed(self, url: str) -> bool:
-        """Проверка, был ли URL уже спарсен"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        cursor.execute("SELECT 1 FROM reviews WHERE url = ? LIMIT 1", (url,))
-        exists = cursor.fetchone() is not None
-
-        conn.close()
-        return exists
-
-    def get_parsing_stats(self) -> Dict:
-        """Получение статистики парсинга"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        # Общая статистика
-        cursor.execute("SELECT COUNT(*) FROM reviews")
-        total_reviews = cursor.fetchone()[0]
-
-        cursor.execute("SELECT COUNT(DISTINCT brand) FROM reviews")
-        unique_brands = cursor.fetchone()[0]
-
-        cursor.execute("SELECT COUNT(DISTINCT model) FROM reviews")
-        unique_models = cursor.fetchone()[0]
-
-        cursor.execute("SELECT source, COUNT(*) FROM reviews GROUP BY source")
-        by_source = dict(cursor.fetchall())
-
-        cursor.execute("SELECT type, COUNT(*) FROM reviews GROUP BY type")
-        by_type = dict(cursor.fetchall())
-
-        conn.close()
-
-        return {
-            "total_reviews": total_reviews,
-            "unique_brands": unique_brands,
-            "unique_models": unique_models,
-            "by_source": by_source,
-            "by_type": by_type,
-        }
-
-
-# ==================== ПАРСЕРЫ ====================
-
-
-from parsers import DromParser, Drive2Parser
-
+from src.config.settings import Config
+from src.database.reviews_database import ReviewsDatabase
+from src.parsers import DromParser, Drive2Parser
+from src.models import ReviewData
 
 # ==================== ГЛАВНЫЙ ПАРСЕР ====================
