+++ auto_reviews_parser/tests/unit/test_review_data.py
@@ -1,27 +1,7 @@
 import hashlib
 from datetime import datetime
-import pathlib
-import re
 
-
-def get_review_data_class():
-    source_path = pathlib.Path(__file__).resolve().parents[1] / "auto_reviews_parser.py"
-    text = source_path.read_text()
-    match = re.search(
-        r"@dataclass\nclass ReviewData:[\s\S]+?def __post_init__\(self\):[\s\S]+?self.content_hash = hashlib.md5\(.+?\)\n",
-        text,
-    )
-    snippet = match.group(0)
-    namespace = {}
-    exec(
-        "from dataclasses import dataclass\nfrom datetime import datetime\nfrom typing import Optional\nimport hashlib\n"
-        + snippet,
-        namespace,
-    )
-    return namespace["ReviewData"]
-
-
-ReviewData = get_review_data_class()
+from src.models import ReviewData
 
 
 def test_parsed_at_auto_set():
